Create writing a c++program to find the exact length of the rope and the exact quantity of carpet that is required? Input format: Input consists of 2 integers.
writing a program to find the exact length of the rope and the exact quantity of carpet that is required? Input format: Input consists of 2 integers. The first integer corresponds to the length of the ground and the second integer corresponds to the breadth of the ground. Output Format: Output Consists of two integers. The first integer corresponds to the length. The second integer corresponds to the quantity of carpet required.