Create Bike road selection algorithm
Joshua is interested very much in seeing Bike race. One Bike is coming so fast. The road on which the Bike comes is divided in to three . There are three roads A, B and C. Each road's distance is given in metres. Before the Bike reaches the end of the road Joshua wants to know that out of three roads in which road the Bike will go if he maintains the given speed. Help him in finding in which road the Bike will travel?

 

INPUT & OUTPUT FORMAT:

Input consists of 4 integers.

First input corresponds to the distance of road A.

Second input corresponds to the distance of road B.

Third input corresponds to the distance of road C.

Fourth input corresponds to the speed of the bike.

If the Bike cannot travel on any road, display “No path exist”.


Constraints:
NA


Example:
Sample Input:

243

136

567

17

Sample Output:

Bike 1 goes into road B