Create Write a function to find the type of array as whether it is even, odd or mixed. If all the elements of an array are even, then display the array type as even. If all the elements of an array are odd, then display the array type as odd. If an array has both even and odd elements, then display the array type as mixed.  If the function returns '1', then print 'Even'. If the function returns '2', then print 'Odd', otherwise print 'Mixed'
Write a function to find the type of array as whether it is even, odd or mixed. If all the elements of an array are even, then display the array type as even. If all the elements of an array are odd, then display the array type as odd. If an array has both even and odd elements, then display the array type as mixed.

If the function returns '1', then print 'Even'.
If the function returns '2', then print 'Odd', otherwise print 'Mixed'.


INPUT FORMAT:

Input consists of 1 integer and 1 array.
Integer corresponds to the size of the array.

Constraints:
FUNCTIONAL REQUIREMENTS:

int findType(int, int *);