Create An upper triangular matrix is a square matrix in which all the elements below the diagonal are zero. That is, all the non-zero elements are in the upper triangle. Write a C program to find whether a given matrix is an upper triangular matrix or not.
An upper triangular matrix is a square matrix in which all the elements below the diagonal are zero. That is, all the non-zero elements are in the upper triangle. Write a C program to find whether a given matrix is an upper triangular matrix or not.

 

 

Input Format:
The input consists of (n*n+1) integers.
The first integer corresponds to the number of rows/columns in the matrix.
The remaining integers correspond to the elements in the matrix.
The elements are read in row wise order, first row first, then second row and so on.
Assume that the maximum value of m and n is 5.

Output Format:
Print yes if it is an upper triangular matrix . Print no if it is not an upper triangular matrix.


Constraints:
NA


Example:
Sample Input 1:
2
1 2
0 3

Sample Output 1:
Yes

Sample Input 2:
2
1 2
3 3

Sample Output 2:
No


Explanation:
NA


Public Test Cases:
#	Input	Expected Output
1	
2
1 2
0 3
Yes
2	
2
1 2
3 3
No