Create A lower triangular matrix is a square matrix in which all the elements above the diagonal are zero. That is, all the non-zero elements are in the lower triangle. Write a C program to find whether a given matrix is a lower triangular matrix or not.
A lower triangular matrix is a square matrix in which all the elements above the diagonal are zero. That is, all the non-zero elements are in the lower triangle. Write a C program to find whether a given matrix is a lower triangular matrix or not.
 
 
Input Format:The input consists of (n*n+1) integers.The first integer corresponds to the number of rows/columns in the matrix.The remaining integers correspond to the elements in the matrix.The elements are read in row wise order, first row first, then second row and so on.Assume that the maximum value of m and n is 5.
Output Format:Print yes if it is a lower triangular matrix . Print no if it is not a lower triangular matrix.

Constraints:

NA

Example:

Sample Input 1:21 02 3
Sample Output 1:yes
Sample Input 2:
21 23 3
Sample Output 2:no

Explanation:

NA

Public Test Cases:
#InputExpected Output1
2 1 2 0 3
no
2
2 1 2 3 3
no