Create Friendly Pair or not.
Write a program to find whether the given numbers are Friendly Pair or not.

Friendly pair is two or more numbers with a common abundancy index i.e the ratio between the sum of divisors of a number and the number itself. For example, 6 and 28 are a Friendly pair as (Sum of divisors of 6)/6 = (Sum of divisors of 28)/28

Note: Don't use unnecessary sentences in input or output. 

 

 INPUT & OUTPUT FORMAT:

Input consists of 2 integers.
The first integer corresponds to number 1 and the second integer corresponds to number 2.
If they are a Friendly pair, display “Friendly Pair” or display “Not Friendly Pair”.

Constraints:
NA


Example:
SAMPLE INPUT: